n64: widen RSP program counter storage
Store RSP IPU PC as 32-bit so JIT word-sized accesses cannot clobber adjacent state. This keeps the fast epilogue path behaviorally stable while preserving 12-bit address masking semantics.

+++ ares/n64/rsp/rsp.hpp
@@ -291,7 +291,7 @@ struct RSP : Thread, Memory::RCP<RSP> {
     };
 
     r32 r[32];
-    u16 pc; // previously u12; now u16 for performance.
+    u32 pc;
   } ipu;
 
   struct Branch {
